requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
mysql-connector-python